Breaking create proxy, write-error out into files, adding tests to create-proxy
