+++ src/write-error.js
@@ -1,0 +1,4 @@
+module.exports = (res, status, code, message) => {
+  res.writeHead(status, { 'Content-Type': 'application/json' });
+  res.end(JSON.stringify({ code, message }));
+};
